chore: move data module to up /src
